Refactor action conversion
