+++ always-succeeds-contract/src/logic/tests.rs
@@ -1,0 +1,1 @@
+use super::*;
