+++ src/backend/tests.rs
@@ -1,5 +1,6 @@
 #![allow(non_snake_case)]
 use super::*;
+use crate::PolicyId;
 use proptest::prelude::*;
 use proptest::test_runner::TestRng;
